code/requirements.txt
Django>=5.0
psycopg2-binary>=2.9
python-dotenv>=1.0
requests>=2.31
openai>=1.0
google-api-python-client
google-auth-httplib2
google-auth-oauthlib